Efficancy Test
Basiclly runs 10,000 trials averaging the final waste and hours of each run. Each run has a random number of 1-50 workers each with a random efficancy of 1-100 and a random number (1-500) of cogOrders with each order having a random number of 1-100 cogs. Changed the baseSim to return a string with the format of "(doublewaste):(intHoursWorked)".